% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicateurs.R
\name{calculer_pib_par_hab}
\alias{calculer_pib_par_hab}
\title{Calculer le PIB par habitant}
\usage{
calculer_pib_par_hab(donnees)
}
\arguments{
\item{donnees}{Un data.frame avec les colonnes \code{gdp} et
\code{population}.}
}
\value{
Le data.frame enrichi d'une colonne \code{gdp_per_capita}.
}
\description{
Ajoute une colonne \code{gdp_per_capita} au data.frame.
}
\examples{
head(calculer_pib_par_hab(world_data))
}
