% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrer.R
\name{filtrer_par_langue}
\alias{filtrer_par_langue}
\title{Filtrer les pays par langue officielle}
\usage{
filtrer_par_langue(donnees, langue)
}
\arguments{
\item{donnees}{Un data.frame contenant une colonne \code{official_language}.}

\item{langue}{Nom de la langue (ex. \code{"French"}, \code{"Arabic"}).
Si \code{NULL}, renvoie tout le dataset.}
}
\value{
Un data.frame filtré.
}
\description{
Filtrer les pays par langue officielle
}
\examples{
filtrer_par_langue(world_data, "French")
}
